otp via sms and some changes in mysql

+++ schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE user_detail (
     uid INT PRIMARY KEY AUTO_INCREMENT,
     fname  VARCHAR(255) NOT NULL,
     lname VARCHAR(255) NOT NULL,
-    phone INT(20) NOT NULL,
+    phone BIGINT(100) NOT NULL,
     email VARCHAR(40) NOT NULL,
     usertype VARCHAR(20) NOT NULL,
     password VARCHAR(255) NOT NULL
